+++ CHANGELOG.md
@@ -4,7 +4,7 @@
     ## **WORK IN PROGRESS**
 -->
 ## **WORK IN PROGRESS**
-* Added "-lf" option to run the lock file update with the force flag
+* Added "-lf" option to run the lock file update with the `--force` flag
 
 ## 3.6.0 (2023-07-03)
 * `git` plugin: Add the `--tagOnly` flag to only create a tag without pushing the commit to the release branch.